+++ Runtime/Triangulation/AssemblyInfo.cs
@@ -1,0 +1,11 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Unity.2D.Animation.Runtime")]
+[assembly: InternalsVisibleTo("Unity.2D.Animation.Editor")]
+[assembly: InternalsVisibleTo("Unity.2D.Animation.Tests.RuntimeTests")]
+[assembly: InternalsVisibleTo("Unity.2D.Animation.Tests.EditorTests")]
+[assembly: InternalsVisibleTo("Unity.2D.IK.Editor")]
+[assembly: InternalsVisibleTo("Unity.2D.IK.Runtime")]
+[assembly: InternalsVisibleTo("Unity.2D.SpriteShape.Runtime")]
+[assembly: InternalsVisibleTo("Unity.2D.PSDImporter.Editor")]
+[assembly: InternalsVisibleTo("Unity.RenderPipelines.Universal.2D.Runtime")]