+++ Samples~/SpriteAtlas/FromResource/LoadFromResource.cs
@@ -3,33 +3,37 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.U2D;
 
-public class LoadFromResource : MonoBehaviour
+namespace UnityEngine.U2D.Common.Samples
 {
-    void OnEnable()
+    public class LoadFromResource : MonoBehaviour
     {
-        SpriteAtlasManager.atlasRequested += RequestLateBindingAtlas;
-        SpriteAtlasManager.atlasRegistered += AtlasRegistered;
-    }
-
-    void OnDisable()
-    {
-        SpriteAtlasManager.atlasRequested -= RequestLateBindingAtlas;
-        SpriteAtlasManager.atlasRegistered -= AtlasRegistered;
-    }
-
-    void RequestLateBindingAtlas(string tag, System.Action<SpriteAtlas> callback)
-    {
-        if (tag == "ResourceAtlas1")
+        void OnEnable()
         {
-            var sa = UnityEngine.Resources.Load<SpriteAtlas>("ResourceAtlas1");
-            callback(sa);
+            SpriteAtlasManager.atlasRequested += RequestLateBindingAtlas;
+            SpriteAtlasManager.atlasRegistered += AtlasRegistered;
         }
-        else
-            Debug.Log("Error: Late binding callback with wrong atlas tag of " + tag);
+
+        void OnDisable()
+        {
+            SpriteAtlasManager.atlasRequested -= RequestLateBindingAtlas;
+            SpriteAtlasManager.atlasRegistered -= AtlasRegistered;
+        }
+
+        void RequestLateBindingAtlas(string tag, System.Action<SpriteAtlas> callback)
+        {
+            if (tag == "ResourceAtlas1")
+            {
+                var sa = UnityEngine.Resources.Load<SpriteAtlas>("ResourceAtlas1");
+                callback(sa);
+            }
+            else
+                Debug.Log("Error: Late binding callback with wrong atlas tag of " + tag);
+        }
+
+        void AtlasRegistered(SpriteAtlas spriteAtlas)
+        {
+            Debug.LogFormat("Registered {0}.", spriteAtlas.name);
+        }
     }
 
-    void AtlasRegistered(SpriteAtlas spriteAtlas)
-    {
-        Debug.LogFormat("Registered {0}.", spriteAtlas.name);
-    }
 }